[TASK] Prevent superfluous COUNT(*) query by Paginator
The query is always done by QueryResultPaginator::getTotalAmountOfItems().
In bigger systems this can be several milliseconds.

We can prevent it in this cases

i) If hidePagination is turned on, then no pagination is needed.
ii) When maximumNumberOfLinks is set (>0) and is less or equal itemsPerPage,
then no pagination would be rendered at all. In this case, the hidePagination setting is
overwritten for the template in order to be *not* breaking.

The change also delays assigning $event->getAssignedValues() to the view. Not
really interesting, but you can see it as a feature to be able to overwrite pagination
configuration by the event.